renaming state info
