+++ yarn-project/types/src/interfaces/index.ts
@@ -1,4 +1,4 @@
-export * from './state_provider.js';
+export * from './state_info_provider.js';
 export * from './hasher.js';
 export * from './aztec-node.js';
 export * from './pxe.js';
