+++ yarn-project/types/src/interfaces/state_info_provider.ts
